+++ android/src/main/kotlin/net/mullvad/mullvadvpn/model/VoucherSubmission.kt
@@ -1,0 +1,3 @@
+package net.mullvad.mullvadvpn.model
+
+data class VoucherSubmission(val timeAdded: Long, val newExpiry: String)