+++ android/src/main/kotlin/net/mullvad/mullvadvpn/ui/WelcomeFragment.kt
@@ -11,6 +11,7 @@ import android.widget.TextView
 import android.widget.Toast
 import kotlinx.coroutines.delay
 import net.mullvad.mullvadvpn.R
+import net.mullvad.mullvadvpn.ui.widget.Button
 import net.mullvad.mullvadvpn.ui.widget.UrlButton
 import net.mullvad.mullvadvpn.util.JobTracker
 import org.joda.time.DateTime
@@ -39,6 +40,12 @@ class WelcomeFragment : ServiceDependentFragment(OnNoService.GoToLaunchScreen) {
 
         view.findViewById<UrlButton>(R.id.buy_credit).apply {
             prepare(daemon, jobTracker)
+        }
+
+        view.findViewById<Button>(R.id.redeem_voucher).apply {
+            setOnClickAction("openRedeemVoucherDialog", jobTracker) {
+                showRedeemVoucherDialog()
+            }
         }
 
         return view
@@ -126,5 +133,13 @@ class WelcomeFragment : ServiceDependentFragment(OnNoService.GoToLaunchScreen) {
         clipboard.primaryClip = clipData
 
         Toast.makeText(context, toastMessage, Toast.LENGTH_SHORT).show()
+    }
+
+    private fun showRedeemVoucherDialog() {
+        val transaction = fragmentManager?.beginTransaction()
+
+        transaction?.addToBackStack(null)
+
+        RedeemVoucherDialogFragment().show(transaction, null)
     }
 }